Put back LLM scope and float to double change in test Automation modules
#rnx

[CL 21828531 by jerome delattre in ue5-main branch]

+++ Engine/Source/Developer/AutomationController/Private/AutomationControllerManager.h
@@ -574,7 +574,7 @@ private:
 		/** The test runners message address */
 		FMessageAddress OwnerMessageAddress;
 		/** The time since we had a ping from the instance*/
-		float LastPingTime;
+		double LastPingTime;
 		/** Is Confirmed to be running */
 		bool bConfirmed;
 	};
